Add more metadata to .desktop files
Some users may search for "SpeechNote" instead of "Speech Note" when
attempting to open the app. Adding the following line ensures that they
find the application in this case, without displaying "SpeechNote"
without a space to the user.

    Keywords=SpeechNote

This is inspired by LibreOffice Calc's Keywords line. Users can search
for "Microsoft Office" and find LibreOffice Calc, because of this
Keywords line:

    Keywords=Accounting;Stats;OpenDocument Spreadsheet;Chart;Microsoft Excel;Microsoft Works;OpenOffice Calc;ods;xls;xlsx;

Also add a "Comment" line, which is also used in search, and may be
displayed in certain cases to the user.